Move the get related function to its own file.
